Replace most uses of name str with Identifier
Replace most of the uses of a string name field with an Identifier
object. This will simplify compiler error messages since every name will
be tagged with a range by default instead of needing to add another
field on the record. Create a Node base interface that all AST nodes
implement, which requires a range.