Fix name resolution method matching
Change the method matcher in the name resolvr to use MethodHandles
rather than reflections-based Executables. This simplifies the code in
the TypeResolver and fixes an issue where the wrong method type can be
selected. Also adjust Main to grab all of the classes in the runtime
folder rather than just the annotated classes.