Move variable idx resolution
Move the variable index resolution from the name resolution step to the
code generation step. It's an implementation detail of the JVM and is
easy to calculate during code generation. It also allows for more
flexibility since we can now create hidden variables during code
generation time without sticking them in the AST.