Use the typed AST for bytecode generation
Only use the tast for the bytecode generation, eliminating the need for
the name and type resolver in that step. This simplifies and decouples
the codegen tests from the actual type checking step. It should
(theoretically) be faster as well since there are much fewer map
lookups. Replace the type variable lookup table in the type resolver
with a mutable type variable to make this transition easier. There is
still more cleanup to do, there are lots of maps around that are no
longer needed.