Add row types
Implement row types for structs. Add a spread operator for literal
structs that allows you to create a new struct type that combines all
the fields provided along with a union of all the fields from the
structs being spread. Add syntax for methods to use row variables in
their type signatures.