Remove TypeNode impl from TypeParameter
Remove the TypeNode implementation from TypeParameter since it doesn't
actually define a new type or describe a type that a variable should
match. Add convenience funcs to convert TypeParameter into a universal
or type variable. Add a level field to TypeVariable instead of concating
the level to the name to make error messages more clear and to clarify
the use of a type variable.