Start converting ByteCodeGenerator tests to E2E
Convert BytecodeGenerator test to e2e test that perform the full
compilation from a source string. This is easier to maintain long term
since it don't rely on the shape of the typed ast nodes. It's also much
simpler to write the tests this way. As part of converting over the
literal tests, fix the parsing and generation of non-int literals. Add
optional type annotations to let exprs as well.