Generate variant tuple constructor funcs
Generate functions to constructo variant tuples on the module
class where the sum type is defined. These functions just delegate to
the actual constructor, however this simplifies the name and type
checking as it just looks like another function defined on the module
rather than needing to special case a call to a constructor.