Lay groudwork for named structs
Create a separate keyword for declaring type aliases from new types.
Simplify the resolution of named type by consolidating the unresolved
named types into a single record. Create a new NamedTypeNode that
replaces the use of BasicTypeNode for named types and will be used for
patterns and struct constructors. Change BasicTypeNode to only refer to
builtin types. Remove a couple unused items in the parser.