Clarify TypeNode name resolution
Rename TypeNameResolver to TypeNodeNameResolver to clarify the intent of
the class and make it less likely to be confused with NamedTypeResolver.
Change various uses of TypeNode and Type to NamedTypeNode and NamedType
to also clarify intent. Move the variant type map from NamedTypeNode
into ModuleScopedNameResolver to ensure TypeNodeNameResolver has a
single responsibility.