Reduce usage of TypeNode
Avoid refering to TypeNodes and instead deal with types directly. This
removes the need to rewrite the types in type nodes during name
resolution when they refer to named types. It also simplifies some of
the code that had to distinguish between structtypenodes and
tupletypenodes since they both turn into struct types.