Return concrete class from struct type class desc
Update the struct type classDesc method to return the concrete class
type instead of StructBase in the case of named structs. Update struct
assignability to only allow named struct assignment from other structs
of the same name. This clarifies the bytecode as well as opens the door
to using invokevirtual in many places instead of invokedynamic. Update
struct construction in various places to include the struct name as a
prefix.