Add run command
Add a command that builds the sources files and then invokes `java` to
run the compiled results. Fix parsing of the existing example programs
by adding a builtin `Array` type instead of using a special syntax. Add
a hack to specifically generate a method called `mainStatic` as a static
method instead of an instance one.