Fix multifile example
Move the stdlib files to a separate directory instead of putting them in
the resource folder. This allows for easier discovery. Add an options
param to the compiler to allow turning compilation with the stdlib off.
Fix the print statement to handle arbitrary reference and primitive
types. Update the multifile main to reflect the changes to std.