Add recursive descent parser
Add a WIP recursive descent parser to replace the antlr based parser.
The main goal is to have a parser that provides usable error message as
opposed to the completely unscrutable antlr ones. So far this includes
a fully working parser that creates a parse tree. The code to converts
the parse tree into the AST is still a work in progresss, but has at
least some basic files working.